Back-End/requirements.txt
asgiref==3.5.2
backports.zoneinfo==0.2.1
dj-database-url==0.5.0
Django==4.0.5
django-cors-headers==3.13.0
django-heroku==0.3.1
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
psycopg2==2.9.3
PyJWT==2.4.0
python-decouple==3.6
pytz==2022.1
sqlparse==0.4.2
certifi==2022.6.15
cloudinary==1.29.0
django-rest-framework==0.1.0
six==1.16.0
urllib3==1.26.9
whitenoise==6.2.0
